requirements.txt
openai
pygithub
requests
streamlit